turn on rule to detect usage of extraneous dependencies
We want to make sure we have all dependencies used in code in our package.json file. We also do not want to use dev dependencies except in tests.

Optional and peer dependencies have meaning for packages only and it also depends on the package. They might rewrite these setting if needed.